backend/api/dataexchange-ws/requirements.txt
fastapi[all]
uvicorn
SQLAlchemy
pymysql
async-generator
async-exit-stack
cryptography
pyyaml
